no more pathcing android jar, this should fix issues with finding the android sdk in some cases

+++ src/main/groovy/me/tatarka/RetrolambdaTask.groovy
@@ -55,9 +55,9 @@ class RetrolambdaTask extends DefaultTask {
         
         def changes = []
         inputs.outOfDate { changes += it }
-
+        
         // Ensure output is cleared if build is not incremental.
-        if (inputs.incremental && !retrolambda.incremental) {
+        if (inputs.incremental && !changes.isEmpty() && !retrolambda.incremental) {
             outputDir.eachFile { it.delete() }
         } else {
             changes.each { change ->
